Refactor main into startup file
